added portion to not download a reference if it is already downloaded

+++ get_refs.sh
@@ -35,12 +35,18 @@ wd=$(pwd)
 function download_refs {
 	cd ${d}
 	while read ref; do
-		id=$( echo "$ref" | cut -d '/' -f 7 )
-		genome_file="${id}_genomic.fna.gz"
-		# NOTE: the below will not work for the gbff file of the marmoratus, we just did it by hand
-		annotation_file="${id}_genomic.gff.gz"
-		rsync --copy-links --times --verbose rsync://"$ref"/"$genome_file" .
-		rsync --copy-links --times --verbose rsync://"$ref"/"$annotation_file" .
+		if [ ! -f $ref  ]; then
+		    echo "File "$ref" doesn't exist. Downloading now."
+		    id=$( echo "$ref" | cut -d '/' -f 7 )
+		    genome_file="${id}_genomic.fna.gz"
+		    # NOTE: the below will not work for the gbff file of the marmoratus, we just did it by hand
+		    annotation_file="${id}_genomic.gff.gz"
+		    rsync --copy-links --times --verbose rsync://"$ref"/"$genome_file" .
+		    rsync --copy-links --times --verbose rsync://"$ref"/"$annotation_file" .
+		else
+    		    echo "File exists"
+		fi
+
 	done<${wd}/${l}
 }
